+++ cmake/FindCLANG_FORMAT.cmake
@@ -7,8 +7,8 @@
 
 find_program(CLANG_FORMAT_EXECUTABLE
              NAMES clang-format
-	           clang-format-10
-		   clang-format-9
+	                 clang-format-10
+		               clang-format-9
                    clang-format-8
                    clang-format-7
                    clang-format-6.0